Add CLI to send/receive UI events
This commit ads the capability for a tocks instance to be controlled
through a script. This uses JSON serialized TocksEvent and TocksUiEvent
items to be observed through the tocks_cli program.

This allows for easier testing of new features and forces improved
decoupling between the UI and the application logic.

An event server/client have been added that rely on unix sockets on
linux, and fall back to less efficient TCP sockets on windows.