Move AudioManager into UI
The concept of UI has changed with the introduction of the event server.
We are fairly close to supporting the client/server running in different
processes.

On top of that audio is really tied to UX. You may want different sounds
depending on which UI you are.

The combination of these two things makes AudioManager feel a lot like a
UI class.

In addition to the main change, the tocks event processing loop has been
moved to QmlUI. This loop has grown to be more complex and deserves a
better home.