Move TocksUiEvent polling loop outside of QTocks
qmetaobject-rs does not do correct reference counting when dispatching a
method call from QML. This means that a long-standing borrow_mut() call
(such as one held during a long async call for example) will break
single mutable reference rules.

With normal queued/threaded code this is not possible. qmetaobject
forces all usages of a QObject to be on the same thread, and usually a
reference can not leak across multiple stages of the Qt event loop.
Using an async function on a QObject is a special case.

This changeset moves the async code outside of the QTocks object to
allow for mutexless usage in a future commit.

The implementation  also had the side effect of keeping the QTocks
object during destruction.